requirements.txt
asgiref==3.7.2
Django==5.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pillow==10.2.0
django-import-export==0.3.1
openpyxl==3.1.2
